--- a/MaxText/layers/llama2.py
+++ b/MaxText/layers/llama2.py
@@ -110,6 +110,7 @@ class LlamaDecoderLayer(nn.Module):
         ar_cache_axis_order=tuple([int(i) for i in cfg.ar_cache_axis_order.split(",")]),
         compute_axis_order=tuple([int(i) for i in cfg.compute_axis_order.split(",")]),
         reshape_q=cfg.reshape_q,
+        kv_quant_axis=cfg.kv_quant_axis,
     )
 
     attention_lnx = attention_layer(