--- a/MaxText/layers/models.py
+++ b/MaxText/layers/models.py
@@ -97,6 +97,7 @@ class DecoderLayer(nn.Module):
         ar_cache_axis_order=tuple([int(i) for i in cfg.ar_cache_axis_order.split(",")]),
         compute_axis_order=tuple([int(i) for i in cfg.compute_axis_order.split(",")]),
         reshape_q=cfg.reshape_q,
+        kv_quant_axis=cfg.kv_quant_axis,
     )
 
     attention_lnx = attention_layer(