--- a/MaxText/layers/quantizations.py
+++ b/MaxText/layers/quantizations.py
@@ -15,10 +15,11 @@
 """Quantization library."""
 
 import functools
+from typing import Optional
 
 from aqt.jax.v2 import config as aqt_config
 from aqt.jax.v2.flax import aqt_flax
-from common_types import Array, Config
+import common_types
 from dataclasses import dataclass
 import flax.linen as nn
 import jax
@@ -26,6 +27,13 @@ import jax.numpy as jnp
 from jax.tree_util import tree_flatten_with_path, tree_unflatten
 
 MAX_INT8 = 127.5
+
+Array = common_types.Array
+Config = common_types.Config
+AxisIdxes = common_types.AxisIdxes
+AxisNames = common_types.AxisNames
+CACHE_HEADS = common_types.CACHE_HEADS
+CACHE_KV = common_types.CACHE_KV
 
 
 @dataclass
@@ -176,9 +184,16 @@ def configure_kv_quantization(config: Config):
   return False if not config.quantize_kvcache else True
 
 
-def quantize_kv(kv: Array, kv_axis: int):
+def quantize_kv(kv: Array, kv_quant_axis: str, axis_names: AxisNames):
   """Quantize key/values stored in kvcache."""
-  scale = jnp.max(jnp.abs(kv), axis=kv_axis, keepdims=True)
+  if kv_quant_axis == "dkv":
+    max_axis_over = axis_names.index(CACHE_KV)
+  elif kv_quant_axis == "heads_and_dkv":
+    max_axis_over = (
+      axis_names.index(CACHE_HEADS),
+      axis_names.index(CACHE_KV)
+    )
+  scale = jnp.max(jnp.abs(kv), axis=max_axis_over, keepdims=True)
   value = jnp.int8(jnp.rint(kv * (MAX_INT8 / scale)))
   return value, scale