@morgandu 2024-06-18
> Looks good, but there is lot of code duplicated from the compute_axis PR. Can you rebase from latest?

It was not ready for review, PR #667 just got merged.

Rebased to main and ready for review now.